+++ scripts/bench.hpp
@@ -693,7 +693,7 @@ bench_result_t bench_nullary(  //
         result.profiled_seconds = running_seconds;
         result.profiled_calls += 1;
         result.profiled_cpu_cycles += cpu_cycles_at_end - cpu_cycles_at_start;
-        result.cpu_cycles_histogram[cpu_cycles_at_end - cpu_cycles_at_start] += 1;
+        result.cpu_cycles_histogram[static_cast<double>(cpu_cycles_at_end - cpu_cycles_at_start)] += 1;
     }
 
     return result;
@@ -764,7 +764,7 @@ bench_result_t bench_unary(    //
         result.profiled_inputs += call_result.inputs_processed;
         result.profiled_calls += 1;
         result.profiled_cpu_cycles += cpu_cycles_at_end - cpu_cycles_at_start;
-        result.cpu_cycles_histogram[cpu_cycles_at_end - cpu_cycles_at_start] += 1;
+        result.cpu_cycles_histogram[static_cast<double>(cpu_cycles_at_end - cpu_cycles_at_start)] += 1;
     });
     result.profiled_seconds = first_call_duration;
     if (first_call_duration >= env.benchmark_seconds) return result;
@@ -792,10 +792,10 @@ bench_result_t bench_unary(    //
 
         result.profiled_seconds = running_seconds;
         result.profiled_cpu_cycles += t4 - t0;
-        result.cpu_cycles_histogram[t1 - t0] += 1;
-        result.cpu_cycles_histogram[t2 - t1] += 1;
-        result.cpu_cycles_histogram[t3 - t2] += 1;
-        result.cpu_cycles_histogram[t4 - t3] += 1;
+        result.cpu_cycles_histogram[static_cast<double>(t1 - t0)] += 1;
+        result.cpu_cycles_histogram[static_cast<double>(t2 - t1)] += 1;
+        result.cpu_cycles_histogram[static_cast<double>(t3 - t2)] += 1;
+        result.cpu_cycles_histogram[static_cast<double>(t4 - t3)] += 1;
     }
 
     result.profiled_seconds += first_call_duration;
